src/bus/ixp465.c: fix -Wmisleading-indentation
White space change. Fixes a compiler warning.
(And this is also a check on my privileges on the SF git repo)

This was the complete compiler warning:

ixp465.c: In function 'ixp465_bus_new':
ixp465.c:85:10: warning: this 'if' clause does not guard... [-Wmisleading-indentation]
          if (bus == NULL)
          ^~
ixp465.c:87:18: note: ...this statement, but the latter is misleadingly indented as if it is guarded by the 'if'
                  part = bus->part;
                  ^~~~